refactor: split trending_view into separate files and add help info
- Split trending_view.rs into 3 separate files with trending_ prefix:
  - trending_language_selection_view.rs: language selection UI
  - trending_repository_selection_view.rs: repository selection UI
  - trending_unified_view.rs: unified two-step UI
- Move TrendingCache to src/cache/ module for better organization
- Add supported languages list to trending command help
- Update imports and module exports accordingly
