+++ src/cache/mod.rs
@@ -1,5 +1,7 @@
 pub mod challenge_cache;
 pub mod gzip_storage;
+pub mod trending_cache;
 
 pub use challenge_cache::{ChallengeCache, CHALLENGE_CACHE};
 pub use gzip_storage::GzipStorage;
+pub use trending_cache::{TrendingCache, TrendingRepository, TRENDING_CACHE};
