+++ src/cli/views/mod.rs
@@ -1,4 +1,6 @@
 pub mod repo_list_view;
 pub mod repo_play_view;
 mod repo_utils;
-pub mod trending_view;
+pub mod trending_language_selection_view;
+pub mod trending_repository_selection_view;
+pub mod trending_unified_view;
