+++ src/cli/views/trending_language_selection_view.rs
@@ -1,0 +1,181 @@
+use crate::ui::colors::Colors;
+use crate::Result;
+use crossterm::{
+    event::{self, Event, KeyCode, KeyEventKind},
+    execute,
+    terminal::{disable_raw_mode, enable_raw_mode, EnterAlternateScreen, LeaveAlternateScreen},
+};
+use ratatui::{
+    backend::CrosstermBackend,
+    layout::{Alignment, Constraint, Direction, Layout},
+    style::{Modifier, Style},
+    text::{Line, Span},
+    widgets::{Block, Borders, List, ListItem, ListState, Padding, Paragraph},
+    Terminal,
+};
+use std::io;
+
+const SUPPORTED_LANGUAGES: &[(&str, &str)] = &[
+    ("C", "C"),
+    ("C#", "C#"),
+    ("C++", "C++"),
+    ("Dart", "Dart"),
+    ("Go", "Go"),
+    ("Haskell", "Haskell"),
+    ("Java", "Java"),
+    ("JavaScript", "JavaScript"),
+    ("Kotlin", "Kotlin"),
+    ("PHP", "PHP"),
+    ("Python", "Python"),
+    ("Ruby", "Ruby"),
+    ("Rust", "Rust"),
+    ("Scala", "Scala"),
+    ("Swift", "Swift"),
+    ("TypeScript", "TypeScript"),
+];
+
+pub fn render_language_selection_ui() -> Result<Option<String>> {
+    // Setup terminal
+    enable_raw_mode()?;
+    let mut stdout = io::stdout();
+    execute!(stdout, EnterAlternateScreen)?;
+    let backend = CrosstermBackend::new(stdout);
+    let mut terminal = Terminal::new(backend)?;
+
+    // State for list selection
+    let mut list_state = ListState::default();
+    list_state.select(Some(0));
+
+    let result: Result<Option<String>> = loop {
+        terminal.draw(|f| {
+            // Add horizontal padding
+            let outer_chunks = Layout::default()
+                .direction(Direction::Horizontal)
+                .constraints([
+                    Constraint::Length(2), // Left padding
+                    Constraint::Min(1),    // Main content
+                    Constraint::Length(2), // Right padding
+                ])
+                .split(f.area());
+
+            let chunks = Layout::default()
+                .direction(Direction::Vertical)
+                .constraints([
+                    Constraint::Length(3), // Header
+                    Constraint::Min(1),    // Language list
+                    Constraint::Length(1), // Controls at bottom
+                ])
+                .split(outer_chunks[1]);
+
+            // Header
+            let header_lines = vec![Line::from(vec![
+                Span::raw("  "), // Left padding
+                Span::styled(
+                    "🌏 Select Programming Language",
+                    Style::default()
+                        .fg(Colors::info())
+                        .add_modifier(Modifier::BOLD),
+                ),
+            ])];
+
+            let header = Paragraph::new(header_lines).block(
+                Block::default()
+                    .borders(Borders::ALL)
+                    .border_style(Style::default().fg(Colors::border()))
+                    .title("GitType - Trending Languages"),
+            );
+            f.render_widget(header, chunks[0]);
+
+            // Language list
+            let items: Vec<ListItem> = SUPPORTED_LANGUAGES
+                .iter()
+                .enumerate()
+                .map(|(i, (display_name, _))| {
+                    let line_spans = vec![
+                        Span::styled(
+                            format!("{:2}. ", i + 1),
+                            Style::default().fg(Colors::text_secondary()),
+                        ),
+                        Span::styled(
+                            format!("{:<20}", display_name),
+                            Style::default()
+                                .fg(Colors::text())
+                                .add_modifier(Modifier::BOLD),
+                        ),
+                    ];
+
+                    ListItem::new(Line::from(line_spans))
+                })
+                .collect();
+
+            let list = List::new(items)
+                .block(
+                    Block::default()
+                        .borders(Borders::ALL)
+                        .border_style(Style::default().fg(Colors::border()))
+                        .title("Programming Languages")
+                        .title_style(
+                            Style::default()
+                                .fg(Colors::text())
+                                .add_modifier(Modifier::BOLD),
+                        )
+                        .padding(Padding::uniform(1)),
+                )
+                .style(Style::default().fg(Colors::text()))
+                .highlight_style(
+                    Style::default()
+                        .bg(Colors::background_secondary())
+                        .add_modifier(Modifier::BOLD),
+                );
+            f.render_stateful_widget(list, chunks[1], &mut list_state);
+
+            // Controls at bottom
+            let controls_line = Line::from(vec![
+                Span::styled("[↑↓/JK]", Style::default().fg(Colors::key_navigation())),
+                Span::styled(" Navigate  ", Style::default().fg(Colors::text())),
+                Span::styled("[SPACE]", Style::default().fg(Colors::key_action())),
+                Span::styled(" Select  ", Style::default().fg(Colors::text())),
+                Span::styled("[ESC]", Style::default().fg(Colors::key_back())),
+                Span::styled(" Return", Style::default().fg(Colors::text())),
+            ]);
+            let controls = Paragraph::new(controls_line).alignment(Alignment::Center);
+            f.render_widget(controls, chunks[2]);
+        })?;
+
+        if let Event::Key(key) = event::read()? {
+            if key.kind == KeyEventKind::Press {
+                match key.code {
+                    KeyCode::Esc => break Ok(None),
+                    KeyCode::Char('j') | KeyCode::Down => {
+                        if let Some(selected) = list_state.selected() {
+                            if selected < SUPPORTED_LANGUAGES.len() - 1 {
+                                list_state.select(Some(selected + 1));
+                            }
+                        }
+                    }
+                    KeyCode::Char('k') | KeyCode::Up => {
+                        if let Some(selected) = list_state.selected() {
+                            if selected > 0 {
+                                list_state.select(Some(selected - 1));
+                            }
+                        }
+                    }
+                    KeyCode::Char(' ') => {
+                        if let Some(selected) = list_state.selected() {
+                            let (_, lang_code) = SUPPORTED_LANGUAGES[selected];
+                            break Ok(Some(lang_code.to_string()));
+                        }
+                    }
+                    _ => {}
+                }
+            }
+        }
+    };
+
+    // Restore terminal
+    disable_raw_mode()?;
+    execute!(terminal.backend_mut(), LeaveAlternateScreen)?;
+    terminal.show_cursor()?;
+
+    result
+}
